Only try to delete meeting if one is scheduled

+++ js/app/controllers/spreedmeetingcontroller.js
@@ -146,6 +146,10 @@ app.controller('SpreedMeetingController', ['$scope', '$http', '$q', 'SpreedMeeti
 	if ($scope.$parent.registerDeleteHook) {
 		$scope.$parent.registerDeleteHook(function() {
 			var token = getRoomToken();
+			if (!token) {
+				// There's nothing to delete
+				return;
+			}
 			SpreedMeetingService.deleteRoom(token);
 		});
 	}
